@@ -8,52 +8,43 @@ class UnicornOrientation {
 
 class UnicornButton extends FloatingActionButton {
   final FloatingActionButton currentButton;
-  final String labelText;
+  final String? labelText;
   final double labelFontSize;
-  final Color labelColor;
-  final Color labelBackgroundColor;
-  final Color labelShadowColor;
+  final Color? labelColor;
+  final Color? labelBackgroundColor;
+  final Color? labelShadowColor;
   final bool labelHasShadow;
   final bool hasLabel;
 
   UnicornButton(
-      {this.currentButton,
-        this.labelText,
-        this.labelFontSize = 14.0,
-        this.labelColor,
-        this.labelBackgroundColor,
-        this.labelShadowColor,
-        this.labelHasShadow = true,
-        this.hasLabel = false})
-      : assert(currentButton != null);
+      {required this.currentButton,
+      this.labelText,
+      this.labelFontSize = 14.0,
+      this.labelColor,
+      this.labelBackgroundColor,
+      this.labelShadowColor,
+      this.labelHasShadow = true,
+      this.hasLabel = false})
+      : super(onPressed: () {});
 
   Widget returnLabel() {
     return Container(
         decoration: BoxDecoration(
             boxShadow: this.labelHasShadow
                 ? [
-              new BoxShadow(
-                color: this.labelShadowColor == null
-                    ? Color.fromRGBO(204, 204, 204, 1.0)
-                    : this.labelShadowColor,
-                blurRadius: 3.0,
-              ),
-            ]
+                    new BoxShadow(
+                      color: this.labelShadowColor == null ? Color.fromRGBO(204, 204, 204, 1.0) : this.labelShadowColor!,
+                      blurRadius: 3.0,
+                    ),
+                  ]
                 : null,
-            color: this.labelBackgroundColor == null
-                ? Colors.white
-                : this.labelBackgroundColor,
+            color: this.labelBackgroundColor == null ? Colors.white : this.labelBackgroundColor,
             borderRadius: BorderRadius.circular(3.0)), //color: Colors.white,
         padding: EdgeInsets.all(9.0),
-        child: Text(this.labelText,
+        child: Text(this.labelText!,
             style: TextStyle(
-                fontSize: this.labelFontSize,
-                fontWeight: FontWeight.bold,
-                color: this.labelColor == null
-                    ? Color.fromRGBO(119, 119, 119, 1.0)
-                    : this.labelColor)));
+                fontSize: this.labelFontSize, fontWeight: FontWeight.bold, color: this.labelColor == null ? Color.fromRGBO(119, 119, 119, 1.0) : this.labelColor)));
   }
-
 
   Widget build(BuildContext context) {
     return this.currentButton;
@@ -63,52 +54,48 @@ class UnicornButton extends FloatingActionButton {
 class UnicornDialer extends StatefulWidget {
   final int orientation;
   final Icon parentButton;
-  final Icon finalButtonIcon;
+  final Icon? finalButtonIcon;
   final bool hasBackground;
-  final Color parentButtonBackground;
-  final List<UnicornButton> childButtons;
+  final Color? parentButtonBackground;
+  final List<UnicornButton>? childButtons;
   final int animationDuration;
   final int mainAnimationDuration;
   final double childPadding;
   final Color backgroundColor;
-  final Function onMainButtonPressed;
+  final Function? onMainButtonPressed;
   final Object parentHeroTag;
   final bool hasNotch;
 
   UnicornDialer(
-      {this.parentButton,
-        this.parentButtonBackground,
-        this.childButtons,
-        this.onMainButtonPressed,
-        this.orientation = 1,
-        this.hasBackground = true,
-        this.backgroundColor = Colors.white30,
-        this.parentHeroTag = "parent",
-        this.finalButtonIcon,
-        this.animationDuration = 180,
-		this.mainAnimationDuration = 200,
-        this.childPadding = 4.0,
-        this.hasNotch = false})
+      {required this.parentButton,
+      this.parentButtonBackground,
+      this.childButtons,
+      this.onMainButtonPressed,
+      this.orientation = 1,
+      this.hasBackground = true,
+      this.backgroundColor = Colors.white30,
+      this.parentHeroTag = "parent",
+      this.finalButtonIcon,
+      this.animationDuration = 180,
+      this.mainAnimationDuration = 200,
+      this.childPadding = 4.0,
+      this.hasNotch = false})
       : assert(parentButton != null);
 
   _UnicornDialer createState() => _UnicornDialer();
 }
 
-class _UnicornDialer extends State<UnicornDialer>
-    with TickerProviderStateMixin {
-  AnimationController _animationController;
-  AnimationController _parentController;
+class _UnicornDialer extends State<UnicornDialer> with TickerProviderStateMixin {
+  late AnimationController _animationController;
+  late AnimationController _parentController;
 
   bool isOpen = false;
 
   @override
   void initState() {
-    this._animationController = AnimationController(
-        vsync: this,
-        duration: Duration(milliseconds: widget.animationDuration));
+    this._animationController = AnimationController(vsync: this, duration: Duration(milliseconds: widget.animationDuration));
 
-    this._parentController =
-        AnimationController(vsync: this, duration: Duration(milliseconds: widget.mainAnimationDuration));
+    this._parentController = AnimationController(vsync: this, duration: Duration(milliseconds: widget.mainAnimationDuration));
 
     super.initState();
   }
@@ -132,8 +119,7 @@ class _UnicornDialer extends State<UnicornDialer>
   Widget build(BuildContext context) {
     this._animationController.reverse();
 
-    var hasChildButtons =
-        widget.childButtons != null && widget.childButtons.length > 0;
+    var hasChildButtons = widget.childButtons != null && widget.childButtons!.length > 0;
 
     if (!this._parentController.isAnimating) {
       if (this._parentController.isCompleted) {
@@ -152,12 +138,9 @@ class _UnicornDialer extends State<UnicornDialer>
 
     var mainFAB = AnimatedBuilder(
         animation: this._parentController,
-        builder: (BuildContext context, Widget child) {
+        builder: (BuildContext context, Widget? child) {
           return Transform(
-              transform: new Matrix4.diagonal3(vector.Vector3(
-                  _parentController.value,
-                  _parentController.value,
-                  _parentController.value)),
+              transform: new Matrix4.diagonal3(vector.Vector3(_parentController.value, _parentController.value, _parentController.value)),
               alignment: FractionalOffset.center,
               child: FloatingActionButton(
                   isExtended: false,
@@ -166,119 +149,96 @@ class _UnicornDialer extends State<UnicornDialer>
                   onPressed: () {
                     mainActionButtonOnPressed();
                     if (widget.onMainButtonPressed != null) {
-                      widget.onMainButtonPressed();
+                      widget.onMainButtonPressed!();
                     }
                   },
                   child: !hasChildButtons
                       ? widget.parentButton
                       : AnimatedBuilder(
-                      animation: this._animationController,
-                      builder: (BuildContext context, Widget child) {
-                        return Transform(
-                          transform: new Matrix4.rotationZ(
-                              this._animationController.value * 0.8),
-                          alignment: FractionalOffset.center,
-                          child: new Icon(
-                              this._animationController.isDismissed
+                          animation: this._animationController,
+                          builder: (BuildContext context, Widget? child) {
+                            return Transform(
+                              transform: new Matrix4.rotationZ(this._animationController.value * 0.8),
+                              alignment: FractionalOffset.center,
+                              child: new Icon(this._animationController.isDismissed
                                   ? widget.parentButton.icon
                                   : widget.finalButtonIcon == null
-                                  ? Icons.close
-                                  : widget.finalButtonIcon.icon),
-                        );
-                      })));
+                                      ? Icons.close
+                                      : widget.finalButtonIcon!.icon),
+                            );
+                          })));
         });
 
     if (hasChildButtons) {
       var mainFloatingButton = AnimatedBuilder(
           animation: this._animationController,
-          builder: (BuildContext context, Widget child) {
-            return Transform.rotate(
-                angle: this._animationController.value * 0.8, child: mainFAB);
+          builder: (BuildContext context, Widget? child) {
+            return Transform.rotate(angle: this._animationController.value * 0.8, child: mainFAB);
           });
 
-      var childButtonsList = widget.childButtons == null ||
-          widget.childButtons.length == 0
-          ? List<Widget>()
-          : List.generate(widget.childButtons.length, (index) {
-        var intervalValue = index == 0
-            ? 0.9
-            : ((widget.childButtons.length - index) /
-            widget.childButtons.length) -
-            0.2;
+      var childButtonsList = widget.childButtons == null || widget.childButtons!.length == 0
+          ? <Widget>[]
+          : List.generate(widget.childButtons!.length, (index) {
+              var intervalValue = index == 0 ? 0.9 : ((widget.childButtons!.length - index) / widget.childButtons!.length) - 0.2;
 
-        intervalValue =
-        intervalValue < 0.0 ? (1 / index) * 0.5 : intervalValue;
+              intervalValue = intervalValue < 0.0 ? (1 / index) * 0.5 : intervalValue;
 
-        var childFAB = FloatingActionButton(
-            onPressed: () {
-              if (widget.childButtons[index].currentButton.onPressed !=
-                  null) {
-                widget.childButtons[index].currentButton.onPressed();
-              }
+              var childFAB = FloatingActionButton(
+                  onPressed: () {
+                    if (widget.childButtons![index].currentButton.onPressed != null) {
+                      widget.childButtons![index].currentButton.onPressed!();
+                    }
 
-              this._animationController.reverse();
-            },
-            child: widget.childButtons[index].currentButton.child,
-            heroTag: widget.childButtons[index].currentButton.heroTag,
-            backgroundColor:
-            widget.childButtons[index].currentButton.backgroundColor,
-            mini: widget.childButtons[index].currentButton.mini,
-            tooltip: widget.childButtons[index].currentButton.tooltip,
-            key: widget.childButtons[index].currentButton.key,
-            elevation: widget.childButtons[index].currentButton.elevation,
-            foregroundColor:
-            widget.childButtons[index].currentButton.foregroundColor,
-            highlightElevation: widget
-                .childButtons[index].currentButton.highlightElevation,
-            isExtended:
-            widget.childButtons[index].currentButton.isExtended,
-            shape: widget.childButtons[index].currentButton.shape);
+                    this._animationController.reverse();
+                  },
+                  child: widget.childButtons![index].currentButton.child,
+                  heroTag: widget.childButtons![index].currentButton.heroTag,
+                  backgroundColor: widget.childButtons![index].currentButton.backgroundColor,
+                  mini: widget.childButtons![index].currentButton.mini,
+                  tooltip: widget.childButtons![index].currentButton.tooltip,
+                  key: widget.childButtons![index].currentButton.key,
+                  elevation: widget.childButtons![index].currentButton.elevation,
+                  foregroundColor: widget.childButtons![index].currentButton.foregroundColor,
+                  highlightElevation: widget.childButtons![index].currentButton.highlightElevation,
+                  isExtended: widget.childButtons![index].currentButton.isExtended,
+                  shape: widget.childButtons![index].currentButton.shape);
 
-        return Positioned(
-          right: widget.orientation == UnicornOrientation.VERTICAL
-              ? widget.childButtons[index].currentButton.mini ? 4.0 : 0.0
-              : ((widget.childButtons.length - index) * 55.0) + 15,
-          bottom: widget.orientation == UnicornOrientation.VERTICAL
-              ? ((widget.childButtons.length - index) * 55.0) + 15
-              : 8.0,
-          child: Row(children: [
-            ScaleTransition(
-                scale: CurvedAnimation(
-                  parent: this._animationController,
-                  curve:
-                  Interval(intervalValue, 1.0, curve: Curves.linear),
-                ),
-                alignment: FractionalOffset.center,
-                child: (!widget.childButtons[index].hasLabel) ||
-                    widget.orientation ==
-                        UnicornOrientation.HORIZONTAL
-                    ? Container()
-                    : Container(
-                    padding:
-                    EdgeInsets.only(right: widget.childPadding),
-                    child: widget.childButtons[index].returnLabel())),
-            ScaleTransition(
-                scale: CurvedAnimation(
-                  parent: this._animationController,
-                  curve:
-                  Interval(intervalValue, 1.0, curve: Curves.linear),
-                ),
-                alignment: FractionalOffset.center,
-                child: childFAB)
-          ]),
-        );
-      });
+              return Positioned(
+                right: widget.orientation == UnicornOrientation.VERTICAL
+                    ? widget.childButtons![index].currentButton.mini
+                        ? 4.0
+                        : 0.0
+                    : ((widget.childButtons!.length - index) * 55.0) + 15,
+                bottom: widget.orientation == UnicornOrientation.VERTICAL ? ((widget.childButtons!.length - index) * 55.0) + 15 : 8.0,
+                child: Row(children: [
+                  ScaleTransition(
+                      scale: CurvedAnimation(
+                        parent: this._animationController,
+                        curve: Interval(intervalValue, 1.0, curve: Curves.linear),
+                      ),
+                      alignment: FractionalOffset.center,
+                      child: (!widget.childButtons![index].hasLabel) || widget.orientation == UnicornOrientation.HORIZONTAL
+                          ? Container()
+                          : Container(padding: EdgeInsets.only(right: widget.childPadding), child: widget.childButtons![index].returnLabel())),
+                  ScaleTransition(
+                      scale: CurvedAnimation(
+                        parent: this._animationController,
+                        curve: Interval(intervalValue, 1.0, curve: Curves.linear),
+                      ),
+                      alignment: FractionalOffset.center,
+                      child: childFAB)
+                ]),
+              );
+            });
 
       var unicornDialWidget = Container(
           margin: widget.hasNotch ? EdgeInsets.only(bottom: 15.0) : null,
           height: double.infinity,
           child: Stack(
-            //fit: StackFit.expand,
+              //fit: StackFit.expand,
               alignment: Alignment.bottomCenter,
               overflow: Overflow.visible,
-              children: childButtonsList.toList()
-                ..add(Positioned(
-                    right: null, bottom: null, child: mainFloatingButton))));
+              children: childButtonsList.toList()..add(Positioned(right: null, bottom: null, child: mainFloatingButton))));
 
       var modal = ScaleTransition(
           scale: CurvedAnimation(
@@ -295,13 +255,7 @@ class _UnicornDialer extends State<UnicornDialer>
               )));
 
       return widget.hasBackground
-          ? Stack(
-          alignment: Alignment.topCenter,
-          overflow: Overflow.visible,
-          children: [
-            Positioned(right: -16.0, bottom: -16.0, child: modal),
-            unicornDialWidget
-          ])
+          ? Stack(alignment: Alignment.topCenter, overflow: Overflow.visible, children: [Positioned(right: -16.0, bottom: -16.0, child: modal), unicornDialWidget])
           : unicornDialWidget;
     }
 
